feat(parent): code review

See merge request web/myopia-pac/myopia-pac-backend!59

+++ myopia-business/business-aggregation/export-biz/src/main/java/com/wupol/myopia/business/aggregation/export/pdf/BaseExportPdfFileService.java
@@ -1,16 +1,19 @@
 package com.wupol.myopia.business.aggregation.export.pdf;
 
 import cn.hutool.core.util.ZipUtil;
+import com.alibaba.fastjson.JSON;
 import com.vistel.Interface.exception.UtilException;
+import com.wupol.myopia.business.aggregation.export.interfaces.ExportFileService;
 import com.wupol.myopia.business.aggregation.export.pdf.domain.ExportCondition;
-import com.wupol.myopia.business.aggregation.export.pdf.interfaces.ExportFileService;
 import com.wupol.myopia.business.core.common.util.S3Utils;
 import com.wupol.myopia.business.core.system.service.NoticeService;
 import lombok.extern.log4j.Log4j2;
 import org.apache.commons.io.FileUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
+import org.springframework.util.StringUtils;
 
 import java.io.File;
 import java.nio.file.Paths;
@@ -22,7 +25,7 @@ import java.util.UUID;
  **/
 @Log4j2
 @Service
-public abstract class BaseExportFileService implements ExportFileService {
+public abstract class BaseExportPdfFileService implements ExportFileService {
 
     @Value("${report.pdf.save-path}")
     public String pdfSavePath;
@@ -38,23 +41,59 @@ public abstract class BaseExportFileService implements ExportFileService {
      * @param exportCondition 导出条件
      * @return void
      **/
+    @Async
     @Override
     public void export(ExportCondition exportCondition) {
-        String fileName = getFileName(exportCondition);
-        String parentPath = getFileSaveParentPath();
-        String fileSavePath = getFileSavePath(parentPath, fileName);
+        String fileName = null;
+        String parentPath = null;
         try {
-            generateFile(exportCondition, fileSavePath, fileName);
+            // 1.获取文件名
+            fileName = getFileName(exportCondition);
+            // 2.获取文件保存父目录路径
+            parentPath = getFileSaveParentPath();
+            // 3.获取文件保存路径
+            String fileSavePath = getFileSavePath(parentPath, fileName);
+            // 4.生成导出的文件
+            generatePdfFile(exportCondition, fileSavePath, fileName);
+            // 5.压缩文件
             File file = compressFile(fileSavePath, fileName);
+            // 6.上传文件
             Integer fileId = uploadFile(file);
+            // 7.发送成功通知
             sendSuccessNotice(exportCondition.getApplyExportFileUserId(), fileName, fileId);
         } catch (Exception e) {
-            log.error("【生成报告异常】{}", fileName, e);
-            sendFailNotice(exportCondition.getApplyExportFileUserId(), fileName);
+            String requestData = JSON.toJSONString(exportCondition);
+            log.error("【生成报告异常】{}", requestData, e);
+            // 发送失败通知
+            if (!StringUtils.isEmpty(fileName)) {
+                sendFailNotice(exportCondition.getApplyExportFileUserId(), fileName);
+            }
         } finally {
+            // 8.删除临时文件
             deleteTempFile(parentPath);
         }
     }
+
+    /**
+     * 导出前的校验
+     *
+     * @param exportCondition 导出条件
+     * @return void
+     **/
+    @Override
+    public void validateBeforeExport(ExportCondition exportCondition) {
+        // 有需要校验的，重写覆盖该方法
+    }
+
+    /**
+     * 生成文件
+     *
+     * @param exportCondition 导出条件
+     * @param fileSavePath 文件保存路径
+     * @param fileName 文件名
+     * @return void
+     **/
+    public abstract void generatePdfFile(ExportCondition exportCondition, String fileSavePath, String fileName);
 
     /**
      * 压缩文件
@@ -63,7 +102,6 @@ public abstract class BaseExportFileService implements ExportFileService {
      * @param fileName 文件名
      * @return java.io.File
      **/
-    @Override
     public File compressFile(String fileSavePath, String fileName) {
         return ZipUtil.zip(fileSavePath);
     }
@@ -113,6 +151,9 @@ public abstract class BaseExportFileService implements ExportFileService {
      **/
     @Override
     public void deleteTempFile(String directoryPath) {
+        if (StringUtils.isEmpty(directoryPath)) {
+            return;
+        }
         FileUtils.deleteQuietly(new File(directoryPath));
     }
 
@@ -121,7 +162,6 @@ public abstract class BaseExportFileService implements ExportFileService {
      *
      * @return java.lang.String
      **/
-    @Override
     public String getFileSaveParentPath() {
         return Paths.get(pdfSavePath, UUID.randomUUID().toString()).toString();
     }
@@ -133,7 +173,6 @@ public abstract class BaseExportFileService implements ExportFileService {
      * @param fileName 文件名
      * @return java.lang.String
      **/
-    @Override
     public String getFileSavePath(String parentPath, String fileName) {
         return Paths.get(parentPath, fileName).toString();
     }
